organised library for p2prc; moved Error, Environment and JSON logic to their own files
