+++ haskell/lib/API.hs
@@ -1,0 +1,4 @@
+
+module API where
+
+a=0
