+++ haskell/lib/CLI.hs
@@ -1,0 +1,3 @@
+module CLI where
+
+a=0
